Switch over to the refactored app
